% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facies_names.R
\name{facies_names}
\alias{facies_names}
\title{extract facies names from stratigraphic column}
\usage{
facies_names(s)
}
\arguments{
\item{s}{a stratcol object}
}
\value{
a vector of facies names for each bed
}
\description{
extract facies names from stratigraphic column
}
\seealso{
\code{\link[=unique_facies_names]{unique_facies_names()}} to get a list of unique facies names
}
