% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base.R
\name{get_base}
\alias{get_base}
\title{find base of column}
\usage{
get_base(s)
}
\arguments{
\item{s}{stratigraphic column}
}
\value{
scalar. position of lowest bed boundary in the stratigraphic column
}
\description{
find base of column
}
