% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_L_unit.R
\name{get_L_unit}
\alias{get_L_unit}
\title{extract length unit from stratigraphic columns}
\usage{
get_L_unit(s)
}
\arguments{
\item{s}{stratigraphic column (a \code{stratcol} object)}
}
\value{
string or NULL, the length unit of the stratigraphic column
}
\description{
extract length unit from stratigraphic columns
}
\examples{
s = as_stratcol(c(0.5, 1, 0.3, 0.7), c("sand", "shale", "sand", "shale"), L_unit = "m")
get_L_unit(s) # returns "m"
}
