% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mom.R
\name{get_mom}
\alias{get_mom}
\title{determine markov order metric by Burgess (2016)}
\usage{
get_mom(m)
}
\arguments{
\item{m}{facies transition matrix}
}
\value{
scalar, the matric order metric
}
\description{
determine markov order metric by Burgess (2016)
}
