% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_facies_names.R
\name{order_facies_names}
\alias{order_facies_names}
\title{order facies names according to appearance}
\usage{
order_facies_names(s)
}
\arguments{
\item{s}{a stratcol object}
}
\value{
a stratcol object
}
\description{
enumeratrs the facies according to their order of appearance (counting from the bottom of the section). To be applied to stratigraphic columns before get_mom is called
}
