% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stratcol.R
\name{plot.stratcol}
\alias{plot.stratcol}
\title{basic plotting of stratigraphic columns}
\usage{
\method{plot}{stratcol}(x, ...)
}
\arguments{
\item{x}{stratigraphic column (a \code{stratcol} object)}

\item{...}{further plotting options. ignored}
}
\value{
invisible NULL
}
\description{
wraps around \code{StratigrapheR::litholog()} to plot a stratigraphic column. The beds are plotted as polygons, the boundaries as horizontal lines.
}
\examples{
s = as_stratcol(c(0.5, 1, 0.3, 0.7), c(1,2,3,1.5), L_unit = "m")
# facies codes are used as hardness
plot(s)
}
