% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stratcol.R
\name{print.stratcol}
\alias{print.stratcol}
\title{print stratigraphic column to console}
\usage{
\method{print}{stratcol}(x, ...)
}
\arguments{
\item{x}{stratigraphic column}

\item{...}{other parameters (ignored)}
}
\value{
invisible NULL, prints to the console
}
\description{
print stratigraphic column to console
}
\seealso{
\code{\link[=summary.stratcol]{summary.stratcol()}} for a summary of a stratigraphic column
}
