% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_facies.R
\name{rename_facies}
\alias{rename_facies}
\title{rename facies}
\usage{
rename_facies(s, new_names, old_names = NULL)
}
\arguments{
\item{s}{stratcol object}

\item{new_names}{new facies names}

\item{old_names}{NULL or a list of old facies names. If NULL, all old facies names will be used}
}
\value{
a stratcol object
}
\description{
replaces old facies names with new ones
}
