% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_thickness.R
\name{total_thickness}
\alias{total_thickness}
\title{get total thickness}
\usage{
total_thickness(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other parameters}
}
\value{
scalar. total thickness
}
\description{
get total thickness
}
