% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_thickness.R
\name{total_thickness.stratcol}
\alias{total_thickness.stratcol}
\title{get total thickness of stratigraphic column}
\usage{
\method{total_thickness}{stratcol}(s, ...)
}
\arguments{
\item{s}{stratigraphic column}

\item{...}{other parameters}
}
\value{
scalar, thickness of column
}
\description{
get total thickness of stratigraphic column
}
