% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_count_matrix.R
\name{trans_count_matrix}
\alias{trans_count_matrix}
\title{facies transition count matrix}
\usage{
trans_count_matrix(x, ...)
}
\arguments{
\item{x}{stratigraphic column}

\item{...}{other parameters. currently ignored}
}
\value{
a transition count matrix of S3 class \code{fa_tran_mat_c}
}
\description{
determines the number of facies transitons in a stratigraphic column and stores the output in a matrix
}
