% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{transition_matrix}
\alias{transition_matrix}
\title{estimate transition matrix from strat column}
\usage{
transition_matrix(s)
}
\arguments{
\item{s}{stratcol object}
}
\value{
a matrix of S3 class fa_tran_mat (facies transition matrix). Has dimension names "from" and "to", and facies as row/column names.
}
\description{
estimate transition matrix from strat column
}
