% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_facies_names.R
\name{unique_facies_names}
\alias{unique_facies_names}
\title{return unique facies names from a stratigraphic column}
\usage{
unique_facies_names(s)
}
\arguments{
\item{s}{a strat col object}
}
\value{
a vector of unique facies names in the stratigraphic column
}
\description{
return unique facies names from a stratigraphic column
}
