% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stratcol.R
\name{as_stratcol}
\alias{as_stratcol}
\title{define stratigraphic column}
\usage{
as_stratcol(thickness, facies, L_unit = NULL, base = 0)
}
\arguments{
\item{thickness}{numeric vector, bed thicknesses}

\item{facies}{vector, facies code of beds}

\item{L_unit}{length unit of bed thickness}

\item{base}{position of lowest bed boundary}
}
\value{
an object of S3 class \code{stratcol}
}
\description{
defines an S3 object \code{stratcol} representing a stratigraphic column. Does not check for the validity of the constructed object. For this, use \code{is_stratcol}
}
\seealso{
\code{\link[=is_stratcol]{is_stratcol()}} to check for validity
}
