% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stratcol.R
\name{as_stratcol}
\alias{as_stratcol}
\title{define stratigraphic columns object}
\usage{
as_stratcol(thickness, facies, L_unit = NULL, base = 0, change_names = TRUE)
}
\arguments{
\item{thickness}{bed thicknesses}

\item{facies}{facies of bed}

\item{L_unit}{SI unit of bed thickenss}

\item{base}{position of lowest bed}

\item{change_names}{should facies names be changed to start with 1?}
}
\description{
defines an S3 object representing a stratigraphic column. Facies names will be changed to start with 1 at the bottom of the column
}
