% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stratcol.R
\name{as_stratcol}
\alias{as_stratcol}
\title{define stratigraphic columns object}
\usage{
as_stratcol(thickness, facies, L_unit = NULL, base = 0)
}
\arguments{
\item{thickness}{bed thicknesses}

\item{facies}{facies of bed}

\item{L_unit}{SI unit of bed thickenss}

\item{base}{position of lowest bed}
}
\description{
defines an S3 object representing a stratigraphic column.
}
\seealso{
\code{\link[=is_stratcol]{is_stratcol()}} to check for validity
}
