% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_thickness.R
\name{bed_thickness}
\alias{bed_thickness}
\title{extract bed thicknesses}
\usage{
bed_thickness(s)
}
\arguments{
\item{s}{a \code{stratcol} object}
}
\value{
a numberic vector of bed thicknesses
}
\description{
extracts bed thicknesses from stratigraphic column
}
