% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facies_names.R
\name{facies_names}
\alias{facies_names}
\title{extract facies names from stratigraphic column}
\usage{
facies_names(s)
}
\arguments{
\item{s}{stratigraphic column (a \code{stratcol} object)}
}
\value{
vector of facies names for each bed
}
\description{
extract facies names from stratigraphic column
}
\examples{
s = as_stratcol(c(0.5, 1, 0.3, 0.7), c("sand", "shale", "sand", "shale"), L_unit = "m")
facies = facies_names(s)
print(facies)

}
\seealso{
\code{\link[=unique_facies_names]{unique_facies_names()}} to get a list of unique facies names
}
