% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facies_repetitions.R
\name{facies_repetitions}
\alias{facies_repetitions}
\title{have successive beds identical facies}
\usage{
facies_repetitions(s)
}
\arguments{
\item{s}{stratigraphic column}
}
\value{
logical. Do two successive beds have the same facies?
}
\description{
have successive beds identical facies
}
