% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_L_unit.R
\name{get_L_unit}
\alias{get_L_unit}
\title{extract length unit from stratigraphic columns}
\usage{
get_L_unit(s)
}
\arguments{
\item{s}{stratigraphic columns}
}
\value{
string or NULL, the lenght unit of the stratigraphic columns
}
\description{
extract length unit from stratigraphic columns
}
