% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stratcol.R
\name{is_stratcol}
\alias{is_stratcol}
\title{is valid stratigraphic column?}
\usage{
is_stratcol(x)
}
\arguments{
\item{x}{object to examine}
}
\description{
examines if x is a valid stratcol object
}
\seealso{
\code{\link[=as_stratcol]{as_stratcol()}} to define stratcol objects
}
