% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_beds.R
\name{merge_beds}
\alias{merge_beds}
\title{merge beds in stratigraphic column}
\usage{
merge_beds(s, mode = "identical facies", ...)
}
\arguments{
\item{s}{stratigraphic column}

\item{mode}{character. criteria for merging. currently onlye "identical facies" is implemented}

\item{...}{other paratmeters. currently not used}
}
\description{
merge beds in stratigraphic column
}
