% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_facies_names.R
\name{order_facies_names}
\alias{order_facies_names}
\title{order facies names according to appearance}
\usage{
order_facies_names(s)
}
\arguments{
\item{s}{stratigraphic column (a \code{stratcol} object)}
}
\value{
a stratigraphic column (a \code{stratcol} object)
}
\description{
enumerates the facies according to their order of appearance (counting from the bottom of the section). To be applied to stratigraphic columns before get_mom is called. Replaces the facies codes by integer numbers
}
\examples{
s = as_stratcol(c(0.5, 1, 0.3, 0.7), c("sand", "shale", "sand", "clay"), L_unit = "m")
s = order_facies_names(s)
plot(s)

}
