% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stratcol.R
\name{plot.stratcol}
\alias{plot.stratcol}
\title{basic plotting of stratigraphic columns}
\usage{
\method{plot}{stratcol}(x, ...)
}
\arguments{
\item{x}{stratcol object}

\item{...}{further plotting options. ignored}
}
\value{
invisible NULL
}
\description{
basic plotting of stratigraphic columns
}
