% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stratcol.R
\name{print.stratcol}
\alias{print.stratcol}
\title{print stratigraphic column to console}
\usage{
\method{print}{stratcol}(x, ...)
}
\arguments{
\item{x}{stratigraphic column (a \code{stratcol} object)}

\item{...}{other parameters (currently ignored)}
}
\value{
invisible NULL, prints to the console
}
\description{
print stratigraphic column to console
}
\examples{
s = as_stratcol(c(0.5, 1, 0.3, 0.7), c("sand", "shale", "sand", "shale"), L_unit = "m")
print(s)

}
\seealso{
\code{\link[=summary.stratcol]{summary.stratcol()}} for a summary of a stratigraphic column
}
