% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_L_unit.R
\name{set_L_unit}
\alias{set_L_unit}
\title{set length unit of strat column}
\usage{
set_L_unit(s, L_unit)
}
\arguments{
\item{s}{stratigraphic column}

\item{L_unit}{string or NULL, the length unit}
}
\description{
set length unit of strat column
}
