% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_col.R
\name{shuffle_col}
\alias{shuffle_col}
\title{rearrange stratigraphic column}
\usage{
shuffle_col(s, allow_rep = TRUE)
}
\arguments{
\item{s}{a stratcol object}

\item{allow_rep}{logical. are repetitions in facies allowed?}
}
\value{
a stratcol object
}
\description{
rearrange stratigraphic column
}
