% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.stratcol.R
\name{summary.stratcol}
\alias{summary.stratcol}
\title{summary of stratigraphic column}
\usage{
\method{summary}{stratcol}(object, ...)
}
\arguments{
\item{object}{stratigraphic column}

\item{...}{further parameters (ignored)}
}
\value{
invisible NULL. prints to the console
}
\description{
summary of stratigraphic column
}
