% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_thickness.R
\name{total_thickness}
\alias{total_thickness}
\title{get total thickness}
\usage{
total_thickness(s, ...)
}
\arguments{
\item{s}{stratigraphic column (a \code{stratcol} object)}

\item{...}{other parameters (currently ignored)}
}
\value{
scalar, total thickness of stratigraphic column
}
\description{
get total thickness
}
\examples{
s = as_stratcol(c(0.5, 1, 0.3, 0.7), c("sand", "shale", "sand", "shale"), L_unit = "m")
total_thickness(s) # returns 2.5


}
