% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{transition_matrix}
\alias{transition_matrix}
\title{estimate transition matrix from strat column}
\usage{
transition_matrix(s)
}
\arguments{
\item{s}{stratcol object}
}
\value{
a matrix
}
\description{
estimate transition matrix from strat column
}
